assistant: Propagate LLM stop reason upwards
