+++ crates/assistant/src/context.rs
@@ -1999,6 +1999,11 @@ impl Context {
                                     });
 
                                 match event {
+                                    LanguageModelCompletionEvent::Stop(reason) => match reason {
+                                        language_model::StopReason::ToolUse => {}
+                                        language_model::StopReason::EndTurn => {}
+                                        language_model::StopReason::MaxTokens => {}
+                                    },
                                     LanguageModelCompletionEvent::Text(chunk) => {
                                         buffer.edit(
                                             [(
